pyproject.toml
[project]
name = "pasteimg-mcp-server"
version = "0.1.0"
description = "Clipboard image capture and share tool"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
]

[dependency-groups]
dev = ["pyright>=1.1.407", "ruff>=0.14.6"]

[project.scripts]
pasteimg-mcp-server = "pasteimg_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "B",  # flake8-bugbear
]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
